+++ src/methods/promise/is.js
@@ -51,7 +51,7 @@ SelfTimer.prototype.is = function(condition) {
       try {
         if (window === "undefined") throw this.messages("Language").onlyBrowser;
 
-        return lang == _helper.__detectLang()
+        return lang.toLowerCase() == _helper.__detectLang()
           ? resolve(true)
           : _Condition === true ? reject(false) : false;
       } catch (e) {
@@ -74,7 +74,7 @@ SelfTimer.prototype.is = function(condition) {
 
         var detect = _helper.__detectLang();
 
-        return lang == detect.slice(0, 2)
+        return lang.toLowerCase() == detect.slice(0, 2)
           ? resolve(true)
           : _Condition === true ? reject(false) : false;
       } catch (e) {
@@ -97,7 +97,10 @@ SelfTimer.prototype.is = function(condition) {
 
         if (!Array.isArray(languages)) throw _message.isNotArray;
 
-        return _helper.__contains(languages, _helper.__detectLang())
+        return _helper.__contains(
+          _helper.__arrayToLower(languages),
+          _helper.__detectLang()
+        )
           ? resolve(true)
           : _Condition === true ? reject(false) : false;
       } catch (e) {
@@ -121,7 +124,10 @@ SelfTimer.prototype.is = function(condition) {
 
       var detect = _helper.__detectLang();
 
-      return _helper.__contains(lang, detect.slice(0, 2))
+      return _helper.__contains(
+        _helper.__arrayToLower(lang),
+        detect.slice(0, 2)
+      )
         ? resolve(true)
         : _Condition === true ? reject(false) : false;
     } catch (e) {
@@ -142,7 +148,10 @@ SelfTimer.prototype.is = function(condition) {
 
       if (!Array.isArray(languages)) throw _message.isNotArray;
 
-      return !_helper.__contains(languages, _helper.__detectLang())
+      return !_helper.__contains(
+        _helper.__arrayToLower(languages),
+        _helper.__detectLang()
+      )
         ? resolve(true)
         : _Condition === true ? reject(false) : false;
     } catch (e) {
@@ -165,7 +174,10 @@ SelfTimer.prototype.is = function(condition) {
 
       var detect = _helper.__detectLang();
 
-      return !_helper.__contains(lang, detect.slice(0, 2))
+      return !_helper.__contains(
+        _helper.__arrayToLower(lang),
+        detect.slice(0, 2)
+      )
         ? resolve(true)
         : _Condition === true ? reject(false) : false;
     } catch (e) {