+++ src/utils/helpers.js
@@ -170,9 +170,18 @@ SelfTimer.prototype.helpers = function() {
    */
   var __detectLang = function() {
     return navigator.userLanguage === "undefined"
-      ? navigator.userLanguage
-      : navigator.language;
+      ? navigator.userLanguage.toLowerCase()
+      : navigator.language.toLowerCase();
   }; // ! __detectLang()
+
+  /**
+   * @return {[ Array ]}
+   */
+  var __arrayToLower = function(arr) {
+    return arr.map(function(val) {
+      return val.toLowerCase();
+    });
+  };
 
   // Register methods
   var REGISTER = {
@@ -184,7 +193,8 @@ SelfTimer.prototype.helpers = function() {
     __dateCompare: __dateCompare,
     __dateString: __dateString,
     __typeToMilliseconds: __typeToMilliseconds,
-    __detectLang: __detectLang
+    __detectLang: __detectLang,
+    __arrayToLower: __arrayToLower
   }; // ! REGISTER
 
   return REGISTER;